Fixed bug with infinite scrolling (Issue #29).
The bug occurred when topics were hidden due to the `Hide SCS` option. When the code tried to get the `BoundingClientRect` of the element, of a hidden topic it would cause the code to break. Now the code checks if the element's `display` is `"none"` and skips it if it is.